added additional MiddleAuthAppCredentialsProvider to fix middle auth app verification and support async request to determine middle auth server
